backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]
requests
aiosqlite